requirements.txt
python-binance
terminaltables
